[WIP] let class definition external
Let class definition external as much as possible.
For example, regard

  class A {
    public:
      int p() {
        return 42;
      }
    private:
      int x;
      std::string q() {
        return "42";
      }
  };

as

  class A {
    public: int p();
    private int x;
    private: std::string q();
  };
  int A::p() {
    return 42;
  }
  std::string q() {
    return "42";
  }

.

The purpose of it is optimization. There is optimizers that modify
classes defined in a XcodeML/C++ program. When modifying class
definition, externally-defined class is easy to handle.

Unfortunately, not all classes can be externally defined easily. So we
deal with only "simple classes."
A class is "XcodeML-simple" when
* the class is POD
* the class doesn't have nested class in it .
This definition is tentative, and can change in future.

We represent XcodeML-simple class definition like this:

    class A {
      public:
        int f() {
          return x;
        }
      private:
        int x;
    };

    <!-- in <typeTable> section... -->
    <classType type="Class0">
      <name>A</name>
      <symbols>
        <id type="Func0" access="public"><name>f</name></id>
        <id type="Func0" access="private"><name>x</name></id>
      </symbols>
    </classType>

    <!-- int <globalDeclarations> section... -->
    <classDecl type="Class0" /> <!-- classDecl is empty -->
    <functionDefinition>
      <name>A::f</name>
      <!-- detail for function definition is described here -->
    </functionDefinition>

When not XcodeML-simple, <classDecl> is not empty.