hide non-nested class decl from <globalDeclarations>
When parsing class declaration with TypeTableVisitor:

First, set normalizability flag of current class "true" if it is not a
inner class (class in the class).
Then parse members recursively.
If encounting inner class, set normalizability flag of the class and
parent class "false."

After TypeTableVisitor, each normalizability flag is correct.