fixup! hide non-nested class decl from <globalDeclarations>
Problem & Solution:
Can't parse classes in namespace.
For example, aborts for segmentation error on parsing below:

    namespace B {
      class C {
        public:
        int x;
      };
    }

isNormalizable decides whether a CXXRecord (not a local class) is
normalizable. In 6ad1678bfb0c4ed4cde3921334a12aaf5e45a8fc, the
implementation storategy is:

    "A class is normalizable iff the parent of it is the
     translation unit."

But this is bad, because a class in namespace is normalizable but its
parent is not the translation unit.
After this misjudgement, PreVisitType lookup the enclosing class of it
but the class doesn't exist, and the program aborts.

This commit fixes the problem. A Class P (not a local) is normalizable
iff a class Q exists and Q encloses P.