Apply the pImpl idiom to `CXXCodeGen::Stream`
Declare `CXXCodeGen::StreamImpl`in Stream.h.

Implementation classes(`StreamImpl`) are usually declared as private
inner classes of the primary class(`Stream`).
But I declared `StreamImpl` as a namespace-scope class because private
inner classes are not able to be passed to non-member functions and it's
inconvenient. I wanted to re-implement `Stream::outputIndentation()` and
`Stream::emit(str)` as non-member functions.