Expand macro to rightly emit `@token`
Update DeclarationsVisitor.cpp.

Use `clang::SourceManager::getSpellingLoc` in order to get the "spelling
location" of a `clang::IntegerLiteral`.  See the previous commit message
(67df24efcdf446d989c0377a10f2585914c7177e) for more information.

A "spelling location" of an expression points to the character sequences that
form the expression.

> The SourceManager can be queried for information about SourceLocation
> objects, turning them into either spelling or expansion locations.
> Spelling locations represent where the bytes corresponding to a token
> came from and expansion locations represent where the location is in
> the user's view.
Ref: https://clang.llvm.org/doxygen/classclang_1_1SourceManager.html

For simplicity, always use spelling locations no matter if an integer
literal is part of a macro expansion or not.