Fix TypeTableInfo::registerType
Problem:
* TypeTableInfo::registerType dereferences nullable pointers.

    const auto RT = dyn_cast<ReferenceType>(T.getTypePtr());
    const auto Pointee = RT->getPointeeType();

* CXXtoXML aborts when the input C/C++ program has a struct declaration
(hereinafter referred to as `struct B`) that has a member of
reference-to-struct-B type, and another struct that is declared before the
declaration of `struct B` (hereinafter referred to as `struct A`) also has a
member of reference-to-struct-B type, e.g.

Solution: See commit 41172f317f5f49b678291afd0be2e00ed73c8a4d.